requirements.txt
python-telegram-bot
binance
requests
httpx
